dont show a default tooltip
Its not needed for the user to understand and it is mostly annoying. If the user opened the config dialog to change the command, theyll see the tooltip command.
Another option is to have the tooltip default to showing the command that was run - that would be useful to users.